Volunteers page and related code added
