+++ layouts/_default/volunteers.html
@@ -1,5 +1,6 @@
 {{ define "main" }}
-    <script type="module" src="/js/volunteers-list.js"></script>
+    <script type="module" src="/js/components/volunteers-list.js"></script>
+    <script type="module" src="/js/components/person-projects.js"></script>
     <volunteers-list>
     {{ .Content }}
     </volunteers-list>
