+++ source/contributors/asf-volunteers.md
@@ -1,5 +1,5 @@
 ---
-title: ASF volunteers (mentors, speakers)
+title: ASF volunteers
 layout: volunteers
 tags: ["committers","newcomers","contributing","mentors","speakers"]
 ---
@@ -9,6 +9,9 @@ and/or to be speakers at ASF-related events.
 
 Project names shown in <strong>bold</strong> mean that the corresponding person
 is a PMC member of that project.
+
+**TODO**: a [number of pages](/tags/speakers.html) explain how to register yourself as a potential speaker.
+These will need to be adapted if we decide to use this page.
 
 ## List of ASF volunteers
